+++ lib/server.dart
@@ -1,13 +1,17 @@
 /// JSON:API server on top of dart:io.
-library server;
+/// WARNING: This library is in beta stage. The API is subject to change.
+library;
 
 export 'package:json_api/src/server/controller.dart';
-export 'package:json_api/src/server/controller_router.dart';
+export 'package:json_api/src/server/cors_middleware.dart';
 export 'package:json_api/src/server/error_converter.dart';
 export 'package:json_api/src/server/errors/collection_not_found.dart';
 export 'package:json_api/src/server/errors/method_not_allowed.dart';
+export 'package:json_api/src/server/errors/not_acceptable.dart';
 export 'package:json_api/src/server/errors/relationship_not_found.dart';
 export 'package:json_api/src/server/errors/resource_not_found.dart';
 export 'package:json_api/src/server/errors/unmatched_target.dart';
+export 'package:json_api/src/server/errors/unsupported_media_type.dart';
+export 'package:json_api/src/server/request_validator.dart';
 export 'package:json_api/src/server/response.dart';
-export 'package:json_api/src/server/try_catch_middleware.dart';
+export 'package:json_api/src/server/router.dart';