+++ lib/src/client/response/relationship_fetched.dart
@@ -1,4 +1,3 @@
-import 'package:http_interop/http_interop.dart' as i;
 import 'package:json_api/document.dart';
 import 'package:json_api/src/client/response.dart';
 
@@ -19,12 +18,6 @@ class RelationshipFetched<R extends Relationship> {
     return RelationshipFetched(response, document.asToOne())
       ..included.addAll(document.included());
   }
-
-  // coverage:ignore-start
-  /// The raw HTTP response
-  @Deprecated('Use rawResponse.httpResponse instead')
-  i.Response get httpResponse => rawResponse.httpResponse;
-  // coverage:ignore-end
 
   /// The raw JSON:API response
   final Response rawResponse;