+++ lib/src/document/error_source.dart
@@ -1,5 +1,7 @@
+import 'package:json_api/src/document/json_encodable.dart';
+
 /// An object containing references to the source of the error.
-class ErrorSource {
+class ErrorSource implements JsonEncodable {
   const ErrorSource({this.pointer = '', this.parameter = ''});
 
   /// A JSON Pointer [RFC6901] to the associated entity in the request document.
@@ -12,6 +14,7 @@ class ErrorSource {
 
   bool get isNotEmpty => !isEmpty;
 
+  @override
   Map<String, String> toJson() => {
         if (parameter.isNotEmpty) 'parameter': parameter,
         if (pointer.isNotEmpty) 'pointer': pointer