+++ lib/src/document/link.dart
@@ -1,6 +1,8 @@
+import 'package:json_api/src/document/json_encodable.dart';
+
 /// A JSON:API link
 /// https://jsonapi.org/format/#document-links
-class Link {
+class Link implements JsonEncodable {
   Link(this.uri);
 
   /// Link URL
@@ -12,6 +14,7 @@ class Link {
   @override
   String toString() => uri.toString();
 
+  @override
   Object toJson() =>
       meta.isEmpty ? uri.toString() : {'href': uri.toString(), 'meta': meta};
 }