+++ lib/src/document/relationship.dart
@@ -1,13 +1,15 @@
 import 'dart:collection';
 
+import 'package:json_api/src/document/json_encodable.dart';
 import 'package:json_api/src/document/link.dart';
 import 'package:json_api/src/document/new_identifier.dart';
 
-class Relationship with IterableMixin<Identifier> {
+class Relationship with IterableMixin<Identifier> implements JsonEncodable {
   final links = <String, Link>{};
   final meta = <String, Object?>{};
 
-  Map<String, dynamic> toJson() => {
+  @override
+  Map<String, Object?> toJson() => {
         if (links.isNotEmpty) 'links': links,
         if (meta.isNotEmpty) 'meta': meta,
       };