+++ lib/src/server/response.dart
@@ -4,14 +4,15 @@ import 'package:json_api/http.dart';
 import 'package:json_api/src/media_type.dart';
 
 /// JSON:API response
-Response response(int statusCode, {OutboundDocument? document}) {
-  final r = Response(
-      statusCode, document != null ? Body.json(document) : Body(), Headers());
-  if (document != null) {
-    r.headers['Content-Type'] = [mediaType];
-  }
-  return r;
-}
+Response response(int statusCode, {OutboundDocument? document}) => Response(
+    statusCode,
+    document != null
+        ? Body.json(document, toEncodable: toJsonEncodable)
+        : Body(),
+    Headers())
+  ..headers.addAll({
+    if (document != null) 'Content-Type': [mediaType]
+  });
 
 Response ok(OutboundDocument document) =>
     response(StatusCode.ok, document: document);
@@ -25,6 +26,9 @@ Response created(OutboundDocument document, String location) =>
 Response notFound([OutboundErrorDocument? document]) =>
     response(StatusCode.notFound, document: document);
 
+Response conflict([OutboundErrorDocument? document]) =>
+    response(StatusCode.conflict, document: document);
+
 Response methodNotAllowed([OutboundErrorDocument? document]) =>
     response(StatusCode.methodNotAllowed, document: document);
 
@@ -34,5 +38,8 @@ Response badRequest([OutboundErrorDocument? document]) =>
 Response unsupportedMediaType([OutboundErrorDocument? document]) =>
     response(StatusCode.unsupportedMediaType, document: document);
 
-Response unacceptable([OutboundErrorDocument? document]) =>
-    response(StatusCode.unacceptable, document: document);
+Response notAcceptable([OutboundErrorDocument? document]) =>
+    response(StatusCode.notAcceptable, document: document);
+
+Response internalServerError([OutboundErrorDocument? document]) =>
+    response(StatusCode.internalServerError, document: document);