+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Automatically encode DateTime objects as an ISO-8601 string
+- The &quot;conflict&quot; server response
+
+### Changed
+- URL Design matching now respects the base URL
+- Allow null to be returned by error interceptors
+
+### Fixed
+- StandardUriDesign working inconsistently, depending on the trailing slash in the path
+
 ## [8.1.0] - 2024-08-29
 ### Added
 - The rawResponse property to the responses of the RoutingClient
@@ -13,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [8.0.0] - 2024-07-01
 ### Added
-- CORS middware
+- CORS middleware
 
 ### Changed
 - Bump http\_interop to v2.0
@@ -264,6 +276,7 @@ the Document model.
 ### Added
 - Client: fetch resources, collections, related resources and relationships
 
+[Unreleased]: https://github.com/f3ath/json-api-dart/compare/8.1.0...HEAD
 [8.1.0]: https://github.com/f3ath/json-api-dart/compare/8.0.0...8.1.0
 [8.0.0]: https://github.com/f3ath/json-api-dart/compare/7.0.1...8.0.0
 [7.0.1]: https://github.com/f3ath/json-api-dart/compare/7.0.0...7.0.1